Make no-std build
