+++ src/lib.rs
@@ -28,12 +28,6 @@ extern crate alloc;
 #[cfg(not(feature = "std"))]
 extern crate core2;
 
-#[cfg(feature = "std")]
-use std::io;
-
-#[cfg(not(feature = "std"))]
-use core2::io;
-
 #[cfg(feature = "serde")]
 #[macro_use]
 extern crate actual_serde as serde;
@@ -52,411 +46,17 @@ pub mod raw;
 pub mod serialize;
 pub mod v0;
 
-use core::fmt;
 #[cfg(feature = "std")]
-use std::collections::{HashMap, HashSet};
+use std::io;
 
-use bitcoin::bip32::{self, KeySource, Xpriv};
-use bitcoin::key::{PrivateKey, PublicKey};
-use bitcoin::secp256k1::{Secp256k1, Signing};
-use bitcoin::{sighash, FeeRate, Transaction};
-
-use crate::error::write_err;
-use crate::prelude::*;
+#[cfg(not(feature = "std"))]
+use core2::io;
 
 #[rustfmt::skip]                // Keep pubic re-exports separate
 pub use crate::{
     error::Error,
     sighash_type::PsbtSighashType,
 };
-
-/// A Partially Signed Transaction.
-type Psbt = v0::Psbt;
-
-/// Data required to call [`GetKey`] to get the private key to sign an input.
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum KeyRequest {
-    /// Request a private key using the associated public key.
-    Pubkey(PublicKey),
-    /// Request a private key using BIP-32 fingerprint and derivation path.
-    Bip32(KeySource),
-}
-
-/// Trait to get a private key from a key request, key is then used to sign an input.
-pub trait GetKey {
-    /// An error occurred while getting the key.
-    type Error: core::fmt::Debug;
-
-    /// Attempts to get the private key for `key_request`.
-    ///
-    /// # Returns
-    /// - `Some(key)` if the key is found.
-    /// - `None` if the key was not found but no error was encountered.
-    /// - `Err` if an error was encountered while looking for the key.
-    fn get_key<C: Signing>(
-        &self,
-        key_request: KeyRequest,
-        secp: &Secp256k1<C>,
-    ) -> Result<Option<PrivateKey>, Self::Error>;
-}
-
-impl GetKey for Xpriv {
-    type Error = GetKeyError;
-
-    fn get_key<C: Signing>(
-        &self,
-        key_request: KeyRequest,
-        secp: &Secp256k1<C>,
-    ) -> Result<Option<PrivateKey>, Self::Error> {
-        match key_request {
-            KeyRequest::Pubkey(_) => Err(GetKeyError::NotSupported),
-            KeyRequest::Bip32((fingerprint, path)) => {
-                let key = if self.fingerprint(secp) == fingerprint {
-                    let k = self.derive_priv(secp, &path)?;
-                    Some(k.to_priv())
-                } else {
-                    None
-                };
-                Ok(key)
-            }
-        }
-    }
-}
-
-/// Map of input index -> pubkey associated with secret key used to create signature for that input.
-pub type SigningKeys = BTreeMap<usize, Vec<PublicKey>>;
-
-/// Map of input index -> the error encountered while attempting to sign that input.
-pub type SigningErrors = BTreeMap<usize, SignError>;
-
-#[rustfmt::skip]
-macro_rules! impl_get_key_for_set {
-    ($set:ident) => {
-
-impl GetKey for $set<Xpriv> {
-    type Error = GetKeyError;
-
-    fn get_key<C: Signing>(
-        &self,
-        key_request: KeyRequest,
-        secp: &Secp256k1<C>
-    ) -> Result<Option<PrivateKey>, Self::Error> {
-        match key_request {
-            KeyRequest::Pubkey(_) => Err(GetKeyError::NotSupported),
-            KeyRequest::Bip32((fingerprint, path)) => {
-                for xpriv in self.iter() {
-                    if xpriv.parent_fingerprint == fingerprint {
-                        let k = xpriv.derive_priv(secp, &path)?;
-                        return Ok(Some(k.to_priv()));
-                    }
-                }
-                Ok(None)
-            }
-        }
-    }
-}}}
-impl_get_key_for_set!(BTreeSet);
-#[cfg(feature = "std")]
-impl_get_key_for_set!(HashSet);
-
-#[rustfmt::skip]
-macro_rules! impl_get_key_for_map {
-    ($map:ident) => {
-
-impl GetKey for $map<PublicKey, PrivateKey> {
-    type Error = GetKeyError;
-
-    fn get_key<C: Signing>(
-        &self,
-        key_request: KeyRequest,
-        _: &Secp256k1<C>,
-    ) -> Result<Option<PrivateKey>, Self::Error> {
-        match key_request {
-            KeyRequest::Pubkey(pk) => Ok(self.get(&pk).cloned()),
-            KeyRequest::Bip32(_) => Err(GetKeyError::NotSupported),
-        }
-    }
-}}}
-impl_get_key_for_map!(BTreeMap);
-#[cfg(feature = "std")]
-impl_get_key_for_map!(HashMap);
-
-/// Errors when getting a key.
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum GetKeyError {
-    /// A bip32 error.
-    Bip32(bip32::Error),
-    /// The GetKey operation is not supported for this key request.
-    NotSupported,
-}
-
-impl fmt::Display for GetKeyError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        use GetKeyError::*;
-
-        match *self {
-            Bip32(ref e) => write_err!(f, "a bip23 error"; e),
-            NotSupported =>
-                f.write_str("the GetKey operation is not supported for this key request"),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for GetKeyError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        use GetKeyError::*;
-
-        match *self {
-            NotSupported => None,
-            Bip32(ref e) => Some(e),
-        }
-    }
-}
-
-impl From<bip32::Error> for GetKeyError {
-    fn from(e: bip32::Error) -> Self { GetKeyError::Bip32(e) }
-}
-
-/// The various output types supported by the Bitcoin network.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[non_exhaustive]
-pub enum OutputType {
-    /// An output of type: pay-to-pubkey or pay-to-pubkey-hash.
-    Bare,
-    /// A pay-to-witness-pubkey-hash output (P2WPKH).
-    Wpkh,
-    /// A pay-to-witness-script-hash output (P2WSH).
-    Wsh,
-    /// A nested segwit output, pay-to-witness-pubkey-hash nested in a pay-to-script-hash.
-    ShWpkh,
-    /// A nested segwit output, pay-to-witness-script-hash nested in a pay-to-script-hash.
-    ShWsh,
-    /// A pay-to-script-hash output excluding wrapped segwit (P2SH).
-    Sh,
-    /// A taproot output (P2TR).
-    Tr,
-}
-
-impl OutputType {
-    /// The signing algorithm used to sign this output type.
-    pub fn signing_algorithm(&self) -> SigningAlgorithm {
-        use OutputType::*;
-
-        match self {
-            Bare | Wpkh | Wsh | ShWpkh | ShWsh | Sh => SigningAlgorithm::Ecdsa,
-            Tr => SigningAlgorithm::Schnorr,
-        }
-    }
-}
-
-/// Signing algorithms supported by the Bitcoin network.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum SigningAlgorithm {
-    /// The Elliptic Curve Digital Signature Algorithm (see [wikipedia]).
-    ///
-    /// [wikipedia]: https://en.wikipedia.org/wiki/Elliptic_Curve_Digital_Signature_Algorithm
-    Ecdsa,
-    /// The Schnorr signature algorithm (see [wikipedia]).
-    ///
-    /// [wikipedia]: https://en.wikipedia.org/wiki/Schnorr_signature
-    Schnorr,
-}
-
-/// Errors encountered while calculating the sighash message.
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum SignError {
-    /// Input index out of bounds.
-    IndexOutOfBounds(IndexOutOfBoundsError),
-    /// Invalid Sighash type.
-    InvalidSighashType,
-    /// Missing input utxo.
-    MissingInputUtxo,
-    /// Missing Redeem script.
-    MissingRedeemScript,
-    /// Missing spending utxo.
-    MissingSpendUtxo,
-    /// Missing witness script.
-    MissingWitnessScript,
-    /// Signing algorithm and key type does not match.
-    MismatchedAlgoKey,
-    /// Attempted to ECDSA sign an non-ECDSA input.
-    NotEcdsa,
-    /// The `scriptPubkey` is not a P2WPKH script.
-    NotWpkh,
-    /// Sighash computation error.
-    SighashComputation(sighash::Error),
-    /// Unable to determine the output type.
-    UnknownOutputType,
-    /// Unable to find key.
-    KeyNotFound,
-    /// Attempt to sign an input with the wrong signing algorithm.
-    WrongSigningAlgorithm,
-    /// Signing request currently unsupported.
-    Unsupported,
-}
-
-impl fmt::Display for SignError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use SignError::*;
-
-        match *self {
-            IndexOutOfBounds(ref e) => write_err!(f, "index out of bounds"; e),
-            InvalidSighashType => write!(f, "invalid sighash type"),
-            MissingInputUtxo => write!(f, "missing input utxo in PBST"),
-            MissingRedeemScript => write!(f, "missing redeem script"),
-            MissingSpendUtxo => write!(f, "missing spend utxo in PSBT"),
-            MissingWitnessScript => write!(f, "missing witness script"),
-            MismatchedAlgoKey => write!(f, "signing algorithm and key type does not match"),
-            NotEcdsa => write!(f, "attempted to ECDSA sign an non-ECDSA input"),
-            NotWpkh => write!(f, "the scriptPubkey is not a P2WPKH script"),
-            SighashComputation(ref e) => write!(f, "sighash: {}", e),
-            UnknownOutputType => write!(f, "unable to determine the output type"),
-            KeyNotFound => write!(f, "unable to find key"),
-            WrongSigningAlgorithm =>
-                write!(f, "attempt to sign an input with the wrong signing algorithm"),
-            Unsupported => write!(f, "signing request currently unsupported"),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for SignError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        use SignError::*;
-
-        match *self {
-            SighashComputation(ref e) => Some(e),
-            IndexOutOfBounds(ref e) => Some(e),
-            InvalidSighashType
-            | MissingInputUtxo
-            | MissingRedeemScript
-            | MissingSpendUtxo
-            | MissingWitnessScript
-            | MismatchedAlgoKey
-            | NotEcdsa
-            | NotWpkh
-            | UnknownOutputType
-            | KeyNotFound
-            | WrongSigningAlgorithm
-            | Unsupported => None,
-        }
-    }
-}
-
-impl From<sighash::Error> for SignError {
-    fn from(e: sighash::Error) -> Self { SignError::SighashComputation(e) }
-}
-
-impl From<IndexOutOfBoundsError> for SignError {
-    fn from(e: IndexOutOfBoundsError) -> Self { SignError::IndexOutOfBounds(e) }
-}
-
-/// This error is returned when extracting a [`Transaction`] from a [`Psbt`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum ExtractTxError {
-    /// The [`FeeRate`] is too high
-    AbsurdFeeRate {
-        /// The [`FeeRate`]
-        fee_rate: FeeRate,
-        /// The extracted [`Transaction`] (use this to ignore the error)
-        tx: Transaction,
-    },
-    /// One or more of the inputs lacks value information (witness_utxo or non_witness_utxo)
-    MissingInputValue {
-        /// The extracted [`Transaction`] (use this to ignore the error)
-        tx: Transaction,
-    },
-    /// Input value is less than Output Value, and the [`Transaction`] would be invalid.
-    SendingTooMuch {
-        /// The original [`Psbt`] is returned untouched.
-        psbt: Psbt,
-    },
-}
-
-impl fmt::Display for ExtractTxError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use ExtractTxError::*;
-
-        match *self {
-            AbsurdFeeRate { fee_rate, .. } =>
-                write!(f, "An absurdly high fee rate of {}", fee_rate),
-            MissingInputValue { .. } => write!(
-                f,
-                "One of the inputs lacked value information (witness_utxo or non_witness_utxo)"
-            ),
-            SendingTooMuch { .. } => write!(
-                f,
-                "Transaction would be invalid due to output value being greater than input value."
-            ),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for ExtractTxError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        use ExtractTxError::*;
-
-        match *self {
-            AbsurdFeeRate { .. } | MissingInputValue { .. } | SendingTooMuch { .. } => None,
-        }
-    }
-}
-
-/// Input index out of bounds (actual index, maximum index allowed).
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum IndexOutOfBoundsError {
-    /// The index is out of bounds for the `psbt.inputs` vector.
-    Inputs {
-        /// Attempted index access.
-        index: usize,
-        /// Length of the PBST inputs vector.
-        length: usize,
-    },
-    /// The index is out of bounds for the `psbt.unsigned_tx.input` vector.
-    TxInput {
-        /// Attempted index access.
-        index: usize,
-        /// Length of the PBST's unsigned transaction input vector.
-        length: usize,
-    },
-}
-
-impl fmt::Display for IndexOutOfBoundsError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use IndexOutOfBoundsError::*;
-
-        match *self {
-            Inputs { ref index, ref length } => write!(
-                f,
-                "index {} is out-of-bounds for PSBT inputs vector length {}",
-                index, length
-            ),
-            TxInput { ref index, ref length } => write!(
-                f,
-                "index {} is out-of-bounds for PSBT unsigned tx input vector length {}",
-                index, length
-            ),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for IndexOutOfBoundsError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        use IndexOutOfBoundsError::*;
-
-        match *self {
-            Inputs { .. } | TxInput { .. } => None,
-        }
-    }
-}
 
 #[cfg(feature = "base64")]
 mod display_from_str {
@@ -466,8 +66,8 @@ mod display_from_str {
     use bitcoin::base64::display::Base64Display;
     use bitcoin::base64::prelude::{Engine as _, BASE64_STANDARD};
 
-    use super::{Error, Psbt};
     use crate::error::write_err;
+    use crate::Error;
 
     /// Error encountered during PSBT decoding from Base64 string.
     #[derive(Debug)]
@@ -502,18 +102,18 @@ mod display_from_str {
         }
     }
 
-    impl Display for Psbt {
+    impl Display for crate::v0::Psbt {
         fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
             write!(f, "{}", Base64Display::new(&self.serialize(), &BASE64_STANDARD))
         }
     }
 
-    impl FromStr for Psbt {
+    impl FromStr for crate::v0::Psbt {
         type Err = PsbtParseError;
 
         fn from_str(s: &str) -> Result<Self, Self::Err> {
             let data = BASE64_STANDARD.decode(s).map_err(PsbtParseError::Base64Encoding)?;
-            Psbt::deserialize(&data).map_err(PsbtParseError::PsbtEncoding)
+            crate::v0::Psbt::deserialize(&data).map_err(PsbtParseError::PsbtEncoding)
         }
     }
 }
